# api-gateway/docs/swagger.yaml
basePath: /
definitions:
  handler.BacktestEstimateRequest:
    properties:
      from_time:
        example: "2024-01-01T00:00:00Z"
        type: string
      strategy:
        example: RSIStrategy
        type: string
      symbol:
        example: BTCUSDT
        type: string
      timeframe:
        example: M1
        type: string
      to_time:
        example: "2024-12-31T23:59:59Z"
        type: string
    type: object
  handler.BacktestRequest:
    properties:
      commission_pct:
        example: 0.001
        type: number
      from_time:
        example: "2024-01-01T00:00:00Z"
        type: string
      initial_capital:
        example: 10000
        type: number
      lot_size:
        example: 0
        type: number
      params:
        additionalProperties: true
        type: object
      script:
        example: // strategy script
        type: string
      slippage_pct:
        example: 0
        type: number
      strategy:
        example: RSIStrategy
        type: string
      symbol:
        example: BTCUSDT
        type: string
      timeframe:
        example: M1
        type: string
      to_time:
        example: "2024-12-31T23:59:59Z"
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Gateway for Mallow algorithmic trading system.

    All protected routes require a Bearer JWT token obtained from the Identity service (`/api/v1/auth/login`).
  title: Mallow API Gateway
  version: "1.0"
paths:
...
  /api/v1/backtest:
    post:
      consumes:
      - application/json
      description: Executes a named strategy backtest over a historical date range.
        Returns a BacktestReport with Sharpe, Sortino, Calmar, max drawdown, win rate,
        profit factor, and trade list.
      parameters:
      - description: Backtest parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.BacktestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Run a backtest
      tags:
      - Herald
  /api/v1/backtest/estimate:
    post:
      consumes:
      - application/json
      description: Returns estimated bar count and time without running the full engine.
      parameters:
      - description: Estimate parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.BacktestEstimateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Estimate backtest cost
      tags:
      - Herald
  /api/v1/backtest/script:
    post:
      consumes:
      - application/json
      description: Executes a script strategy. Always saves the script version and
        result to the store. Pass strategy_id to link to an existing version chain.
      parameters:
      - description: Script backtest parameters (use 'script' field)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.BacktestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Run a script backtest
      tags:
      - Herald
...
  /api/v1/strategies:
    get:
      description: Returns all ~80 named strategy IDs that herald supports.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - BearerAuth: []
      summary: List available strategies
      tags:
      - Herald
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"